requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.1                       # passlib 1.7.4 incompatible with bcrypt 4.1+
python-multipart==0.0.9
aiofiles==24.1.0
psutil==6.0.0
